refactor: デッドコード削除と依存方向違反の修正
- reviewListService.ts + テストを削除（importer ゼロ。v5 の reviewService に置換済み）
- PracticePageLayout を components/ → features/practice/components/ に移動
  （components → features の依存方向違反を解消。AdminLayout と同じ配置パターン）
- 利用5ページとテストモックの import パスを更新

Closes #296
