+++ apps/web/src/pages/MiniProjectsPage.tsx
@@ -5,7 +5,7 @@ import { useDocumentTitle } from '../hooks/useDocumentTitle'
 import { getProjectProgressMap } from '../services/miniProjectService'
 import { PracticeModeNav } from '../features/daily/components/PracticeModeNav'
 import { ProjectCard } from '../features/mini-projects/components/ProjectCard'
-import { PracticePageLayout } from '../components/PracticePageLayout'
+import { PracticePageLayout } from '../features/practice/components/PracticePageLayout'
 import { Spinner } from '../components/Spinner'
 import { MINI_PROJECTS } from '../content/mini-projects/projects'
 import type { MiniProjectProgress } from '../content/mini-projects/types'
