+++ apps/web/src/pages/__tests__/CodeDoctorPage.test.tsx
@@ -21,7 +21,7 @@ vi.mock('@/services/codeDoctorService', () => ({
   submitDoctorSolution: vi.fn(),
 }))
 
-vi.mock('@/components/PracticePageLayout', () => ({
+vi.mock('@/features/practice/components/PracticePageLayout', () => ({
   PracticePageLayout: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
 }))
